Consolidation after testing of the append function
The append function works for the standard and parallel operation. It was tested with a single FPGA and with multiple FPGAs. In parallel operation it is possible to append to only one FPGA.
It is unclear what will be the bahaviour when trying to append a standard schedule to parallel operation and vice versa. For now, it is up to the user to make sure that they are using the append correctly